requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
black==22.8.0
cachetools==4.2.2
certifi==2022.6.15.1
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
PyPDF2==2.10.6
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tzdata==2022.2
tzlocal==4.2
